(EXPERIMENTAL:) Show cluster IDs in verse view
Please note that the cluster IDs change unpredictably when
the clustering is recomputed. The stable way to refer to clusters
is still by referring to a concrete verse, like in the URL of
the cluster view.

This change might need further discussion / testing and might
be reverted in the future.