Perform database checks on startup
- on startup, check which tables are available in the database,
- if a table is not found, disable the related functionality
  and print a warning, but keep working if possible

Closes #37.

There might still be links present that lead to views that crash
(e.g. verse clusters). Such links should be removed, but this can be
handled later. Also the table of contents on the main page has some
hardcoded logic that should be reworked (see #32).

The minimal working set of tables is: `poems`, `verse_poem`, `verses`.