Load the starting page content from the DB
This replaces the hardcoded content of the starting page with blocks
loaded from the database. The table `runoregi_pages` contains blocks
with HTML content (and optional helptext shown next to the title).

If the table is not found or empty, a hardcoded default set of blocks
is shown.

Part of #32